move cart save wearleveling/crypto to separate file
