+++ arm9/source/gamecart/gamecart.c
@@ -1,9 +1,4 @@
 #include "gamecart.h"
-#include "aes.h"
-#include "crc16.h"
-#include "fsdrive.h"
-#include "fsinit.h"
-#include "image.h"
 #include "protocol.h"
 #include "command_ctr.h"
 #include "command_ntr.h"
@@ -12,10 +7,44 @@
 #include "ncch.h"
 #include "ncsd.h"
 #include "rtc.h"
-#include "sha.h"
-#include "unittype.h"
 
 #define CART_INSERTED (!(REG_CARDSTATUS & 0x1))
+
+typedef struct {
+    NcsdHeader ncsd;
+    u32 card2_offset;
+    u8  cinfo0[0x312 - (0x200 + sizeof(u32))];
+    u32 rom_version;
+    u8  cinfo1[0x1000 - (0x312 + sizeof(u32))];
+    NcchHeader ncch;
+    u8  padding[0x3000 - 0x200];
+    u8  private[PRIV_HDR_SIZE];
+    u8  unused[0x4000 + 0x8000 - PRIV_HDR_SIZE]; // 0xFF
+    u32 cart_type;
+    u32 cart_id;
+    u32 cart_id2;
+    u64 cart_size;
+    u64 data_size;
+    u32 save_size;
+    CardSPIType save_type;
+    u32 unused_offset;
+} PACKED_ALIGN(16) CartDataCtr;
+
+typedef struct {
+    TwlHeader ntr_header;
+    u8 ntr_padding[0x3000]; // 0x00
+    u8 secure_area[0x4000];
+    u8 secure_area_enc[0x4000];
+    u8 modcrypt_area[0x4000];
+    u32 cart_type;
+    u32 cart_id;
+    u32 cart_id2; // meaningless on TWL
+    u64 cart_size;
+    u64 data_size;
+    u32 save_size;
+    CardSPIType save_type;
+    u32 arm9i_rom_offset;
+} PACKED_ALIGN(16) CartDataNtrTwl;
 
 static DsTime init_time;
 static bool encrypted_sa = false;
@@ -132,245 +161,6 @@ static u32 GetCtrCartSaveSize(CartData* cdata) {
     }
 }
 
-void SetBlockmapEntry(u32 index, SaveBlockmapEntry *in_entry, CartWearLevelingData *wldata) {
-    u8 *base_ptr = &wldata->blockmap[sizeof(SaveBlockmapHeader)];
-    if (wldata->type == CARD_SAVE_WEAR_LEVELING_V2) {
-        SaveBlockmapEntryV2 *entry = &((SaveBlockmapEntryV2 *)base_ptr)[index];
-        entry->phys_sec = in_entry->phys_sec;
-        entry->remap_count = in_entry->remap_count;
-        entry->used = in_entry->used;
-    } else {
-        SaveBlockmapEntryV1 *entry = &((SaveBlockmapEntryV1 *)base_ptr)[index];
-        entry->phys_sec = in_entry->phys_sec;
-        entry->remap_count = in_entry->remap_count;
-        entry->used = in_entry->used;
-        memcpy(entry->checksums, in_entry->checksums, sizeof(entry->checksums));
-    }
-}
-
-SaveBlockmapEntry GetBlockmapEntry(u32 index, CartWearLevelingData *wldata) {
-    u8 *base_ptr = &wldata->blockmap[sizeof(SaveBlockmapHeader)];
-    if (wldata->type == CARD_SAVE_WEAR_LEVELING_V2) {
-        SaveBlockmapEntryV2 *entries = (SaveBlockmapEntryV2 *)base_ptr;
-        return (SaveBlockmapEntry) { .checksums = { 0 }, .phys_sec = entries[index].phys_sec, .remap_count = entries[index].remap_count, .used = entries[index].used };
-    } else {
-        SaveBlockmapEntryV1 *entries = (SaveBlockmapEntryV1 *)base_ptr;
-        SaveBlockmapEntry out = (SaveBlockmapEntry) { .phys_sec = entries[index].phys_sec, .remap_count = entries[index].remap_count, .used = entries[index].used };
-        memcpy(out.checksums, entries[index].checksums, sizeof(out.checksums));
-        return out;
-    }
-}
-
-void ApplyJournalEntryToBlockmap(SaveJournalEntry *entry, CartWearLevelingData *wldata) {
-    SaveBlockmapEntry ent = {
-        .used = 1,
-        .phys_sec = entry->main.dst_phys_sec,
-        .remap_count = entry->main.new_dst_remap_count
-    };
-    memcpy(ent.checksums, entry->main.new_dst_checksums, sizeof(ent.checksums));
-    
-    // remap the virtual sector to the new physical sector
-    SetBlockmapEntry(entry->main.src_virt_sec, &ent, wldata);
-    
-    // unless an actual remap didn't occur, we must mark (the physical sector
-    // whose virtual sector number we used for remapping this virtual sector to)
-    // as free and ready to use for another remap
-    if (entry->main.dst_phys_sec != entry->main.src_phys_sec) {
-        ent.used = 0;
-        ent.phys_sec = entry->main.src_phys_sec;
-        ent.remap_count = entry->main.src_remap_count;
-        memset(ent.checksums, 0, sizeof(ent.checksums));
-        SetBlockmapEntry(entry->main.dst_virt_sec, &ent, wldata);
-    }
-}
-
-void InitCtrCardSaveCryptoKey(NcsdHeader *ncsd, CartData *cdata) {
-    CartDataCtr *ctr_cdata = (CartDataCtr *)cdata;
-    
-    // save data crypto (if supported)
-    u32 save_media_old = ncsd->partition_flags[7];
-    u32 save_media_new = ncsd->partition_flags[3];
-    u8 save_crypto_keysel_base = ncsd->partition_flags[1];
-    u8 save_crypto_keysel_extra = ncsd->extra_save_keysel;
-    s32 save_crypto_keysel = 0;
-    bool is_card2 = cdata->cart_id & 0x8000000;
-    
-    bool supported_save_crypto =
-        (save_media_old == 0 && save_media_new == 0) || /* SPI save flash (very old carts) */
-        (save_media_old == 1) || /* old SPI flash */
-        (save_media_old == 0 && save_media_new == 1) || /* newer SPI flash */
-        is_card2 /* CARD2 */;
-    
-    // skip everything if there isn't save data or its crypto is not supported
-    if (cdata->save_type == CARD_SAVE_NONE || !supported_save_crypto) {
-        ctr_cdata->save_crypto_type = CARD_SAVE_CRYPTO_INVALID;
-        return;
-    }
-    
-    if (save_media_old == 0 && save_media_new == 0 && !is_card2)
-        ctr_cdata->save_crypto_repeating_ctr = true;
-
-    if (is_card2)
-        save_crypto_keysel = save_crypto_keysel_base + save_crypto_keysel_extra;
-    else if (save_media_old != 0)
-        save_crypto_keysel = 0;
-    else if (save_media_new != 0)
-        save_crypto_keysel = save_crypto_keysel_base + save_crypto_keysel_extra;
-    else
-        save_crypto_keysel = -1;
-    
-    save_crypto_keysel += 1;
-    
-    u8 save_key_y[16] = { 0 };
-    
-    struct {
-        NcchHeader ncch;
-        NcchExtHeader exthdr;
-    } hdrs;
-    
-    ReadCartBytes(&hdrs, ncsd->partitions[0].offset * NCSD_MEDIA_UNIT, sizeof(hdrs), cdata, false);
-    DecryptNcch(&hdrs.exthdr, NCCH_EXTHDR_OFFSET, sizeof(NcchExtHeader), &hdrs.ncch, NULL);
-    u8 *cart_unique_id = cdata->header + 0x4000;
-
-    switch (save_crypto_keysel) {
-        case 0: // "backup security version" -1
-            memcpy(save_key_y, hdrs.exthdr.signature, 8);
-            memcpy(&save_key_y[8], &cdata->cart_id, 4);
-            memcpy(&save_key_y[12], &cdata->cart_id2, 4);
-            ctr_cdata->save_crypto_type = CARD_SAVE_CRYPTO_V0;
-            break;
-
-        case 1: // "backup security version" 0
-        case 11: // "backup security version" 10
-        {
-            // this type is not supported on O3DS
-            if (IS_O3DS && save_crypto_keysel == 11) {
-                ctr_cdata->save_crypto_type = CARD_SAVE_CRYPTO_INVALID;
-                return;
-            }
-
-            u8 tmpbuf[0x48];
-            u8 hash[0x20];
-            memcpy(tmpbuf, hdrs.exthdr.signature, 8);
-            memcpy(&tmpbuf[8], cart_unique_id, 0x40);
-
-            sha_quick(hash, tmpbuf, 0x48, SHA256_MODE);
-            memcpy(save_key_y, hash, 16);
-            ctr_cdata->save_crypto_type = save_crypto_keysel == 11 ? CARD_SAVE_CRYPTO_V1_N3DS : CARD_SAVE_CRYPTO_V1;
-            break;
-        }
-
-        case 2: // "backup security version" 1
-        {
-            static bool save60KeyYSetup = false;
-            
-            if (!save60KeyYSetup) {
-                static const u8 save60KeyY[16] = { 0xC3, 0x69, 0xBA, 0xA2, 0x1E, 0x18, 0x8A, 0x88, 0xA9, 0xAA, 0x94, 0xE5, 0x50, 0x6A, 0x9F, 0x16 };
-                setup_aeskeyY(0x2F, save60KeyY);
-                save60KeyYSetup = true;
-            }
-            
-            u8 hash[0x20];
-            u8 tmpbuf[0x70];
-            
-            u32 in_ncch_offset = hdrs.ncch.offset_exefs * NCCH_MEDIA_UNIT + 0x1E0;
-            ReadCartBytes(&hash, ncsd->partitions[0].offset * NCSD_MEDIA_UNIT + in_ncch_offset, sizeof(hash), cdata, false);
-            DecryptNcch(hash, in_ncch_offset, sizeof(hash), &hdrs.ncch, NULL);
-            
-            memcpy(tmpbuf, hdrs.exthdr.signature, 8);
-            memcpy(&tmpbuf[8], cart_unique_id, 0x40);
-            memcpy(&tmpbuf[0x48], &hdrs.ncch.programId, 8);
-            memcpy(&tmpbuf[0x50], hash, 0x20);
-            sha_quick(hash, tmpbuf, 0x70, SHA256_MODE);
-            
-            use_aeskey(0x2F);
-            aes_cmac(hash, save_key_y, 2);
-            ctr_cdata->save_crypto_type = CARD_SAVE_CRYPTO_V2;
-            break;
-        }
-        default:
-            ctr_cdata->save_crypto_type = CARD_SAVE_CRYPTO_INVALID;
-            return;
-    };
-    
-    u32 cmac_keyslot = ctr_cdata->save_crypto_type == CARD_SAVE_CRYPTO_V1_N3DS ? CARD_SAVE_CMAC_KEYSLOT_N3DS : CARD_SAVE_CMAC_KEYSLOT_O3DS;
-    u32 crypto_keyslot = ctr_cdata->save_crypto_type == CARD_SAVE_CRYPTO_V1_N3DS ? CARD_SAVE_CRYPTO_KEYSLOT_N3DS : CARD_SAVE_CRYPTO_KEYSLOT_O3DS;
-
-    setup_aeskeyY(cmac_keyslot, save_key_y); // savedata CMAC key
-    setup_aeskeyY(crypto_keyslot, save_key_y); // savedata crypto key
-}
-
-void InitCtrCardSaveWearLeveling(CartData *cdata) {
-    CartDataCtr *ctr_cdata = (CartDataCtr *)cdata;
-    ctr_cdata->wear_leveling.initialized = false;
-    
-    // CARD2 does not implement wear leveling for the writable portion of the "ROM"
-    if (ctr_cdata->cart_id & 0x8000000) {
-        ctr_cdata->wear_leveling.type = CARD_SAVE_WEAR_LEVELING_NONE;
-        ctr_cdata->wear_leveling.initialized = true;
-        return;
-    }
-    
-    // SPI flash savedata implements wear leveling
-    u32 num_physical_sectors = ctr_cdata->save_size / 0x1000;
-    ctr_cdata->wear_leveling.logical_sectors = num_physical_sectors - 1;
-    ctr_cdata->wear_leveling.logical_size = (num_physical_sectors - 1) * 0x1000;
-    
-    if (num_physical_sectors > 128) {
-        // V2, header + blockmap_entries_v2[511] + crc
-        ctr_cdata->wear_leveling.blockmap_size = 0x400; // fixed size
-        ctr_cdata->wear_leveling.type = CARD_SAVE_WEAR_LEVELING_V2;
-    } else {
-        // V1, header + blockmap_entries_v1[n] + crc
-        ctr_cdata->wear_leveling.blockmap_size = sizeof(SaveBlockmapHeader) + ctr_cdata->wear_leveling.logical_sectors * sizeof(SaveBlockmapEntryV1) + sizeof(u16);
-        ctr_cdata->wear_leveling.type = CARD_SAVE_WEAR_LEVELING_V1;
-    }
-    
-    // TODO: add CMAC verification ability for DISA 
-    
-    if (ReadCartSave(ctr_cdata->wear_leveling.blockmap, 0, ctr_cdata->wear_leveling.blockmap_size, cdata) != 0)
-        return;
-    
-    u32 journal_size = 0x1000 - ctr_cdata->wear_leveling.blockmap_size;
-    ctr_cdata->wear_leveling.num_journal_entries = journal_size / sizeof(SaveJournalEntry);
-    
-    if (ReadCartSave((u8 *)ctr_cdata->wear_leveling.journal, ctr_cdata->wear_leveling.blockmap_size, journal_size, cdata) != 0)
-        return;
-
-    
-    u8 *crcLoc = &ctr_cdata->wear_leveling.blockmap[ctr_cdata->wear_leveling.blockmap_size - 2];
-    u16 expected_bmap_crc = crcLoc[0] | crcLoc[1] << 8;
-    u16 calc_bmap_crc = crc16_quick(ctr_cdata->wear_leveling.blockmap, ctr_cdata->wear_leveling.blockmap_size - 2);
-    if (expected_bmap_crc != calc_bmap_crc)
-        return;
-    
-    u32 journal_idx = 0;
-
-    for (; journal_idx < ctr_cdata->wear_leveling.num_journal_entries; journal_idx++) {
-        SaveJournalEntry *ent = &ctr_cdata->wear_leveling.journal[journal_idx];
-        
-        // journal entry should point to a valid physical sector
-        if (ent->main.dst_phys_sec) {
-            // reached the end of the journal
-            if (ent->main.src_virt_sec >= ctr_cdata->wear_leveling.logical_sectors)
-                break;
-            
-            ApplyJournalEntryToBlockmap(ent, &ctr_cdata->wear_leveling);
-        }
-    }
-    
-    static const u8 blank[sizeof(SaveJournalEntry)] = { 0xFF };
-    u32 num_bad_journal_entries = 0;
-    for (; journal_idx < ctr_cdata->wear_leveling.num_journal_entries; journal_idx++) {
-        num_bad_journal_entries += memcmp((u8 *)&ctr_cdata->wear_leveling.journal[journal_idx], blank, sizeof(SaveJournalEntry)) == 0;
-    }
-    
-    if (num_bad_journal_entries)
-        return;
-    
-    ctr_cdata->wear_leveling.initialized = true;
-}
-
 u32 InitCartRead(CartData* cdata) {
     get_dstime(&init_time);
     encrypted_sa = false;
@@ -433,7 +223,7 @@ u32 InitCartRead(CartData* cdata) {
 
         bool is_card2 = cdata->cart_id & 0x8000000;
         u32 card2_offset = getle32(cdata->header + 0x200);
-        
+
         if (is_card2 && card2_offset != 0xFFFFFFFF) {
             cdata->save_type = CARD_SAVE_CARD2;
             cdata->save_size = GetCtrCartSaveSize(cdata);
@@ -456,15 +246,6 @@ u32 InitCartRead(CartData* cdata) {
                 cdata->save_size = CardSPIGetCapacity(cdata->spi_save_type);
             }
         }
-        
-        // for compatibility purposes save crypto and wear leveling init are optional
-        
-        // all CTRCARDs use save crypto
-        InitCtrCardSaveCryptoKey(ncsd, cdata);
-        
-        // all CTRCARDs except CARD2 use wear leveling for save data
-        InitCtrCardSaveWearLeveling(cdata);
-
     } else { // NTR/TWL cartridges
         // NTR header
         TwlHeader* nds_header = (void*)cdata->header;
@@ -688,147 +469,8 @@ u32 ReadCartSave(u8* buffer, u64 offset, u64 count, CartData* cdata) {
     }
 }
 
-static u32 ReadDecryptedCard1Save(u8 *buffer, u64 offset, u64 count, CartData* cdata) {
-    CartDataCtr *ctr_cdata = (CartDataCtr *)cdata;
-
-    if (offset >= ctr_cdata->wear_leveling.logical_size) return 1;
-    if (offset + count > ctr_cdata->wear_leveling.logical_size) return 1;
-
-    u32 first_sector = offset / 0x1000;
-    u32 last_sector = (offset + count - 1) / 0x1000;
-    u32 outbuf_offset = 0;
-    
-    SaveBlockmapEntry ent;
-    u8 sectorbuf[0x1000];
-    u8 ctr[16];
-    memset(ctr, 0, sizeof(ctr));
-
-    u32 crypto_keyslot = ctr_cdata->save_crypto_type == CARD_SAVE_CRYPTO_V1_N3DS ? CARD_SAVE_CRYPTO_KEYSLOT_N3DS : CARD_SAVE_CRYPTO_KEYSLOT_O3DS;
-    use_aeskey(crypto_keyslot);
-    
-    // the minimum one can read from the flash is 4K sectors anyway, and blockmap scatters it, so we do it sector by sector
-    for (u32 cur_sector = first_sector; cur_sector < last_sector + 1; cur_sector++) {
-        ent = GetBlockmapEntry(cur_sector, &ctr_cdata->wear_leveling);
-        
-        // where in this sector we need to start reading data from
-        u32 in_sector_start = cur_sector == first_sector ? offset % 0x1000 : 0;
-        // how much data we can read in this sector assuming the start offset above
-        u32 in_sector_size = 0x1000 - in_sector_start;
-        // how much data we actually need to read from the sector
-        u32 chunksize = min(count, in_sector_size);
-        
-        if (!ent.used) {
-            memset(&sectorbuf, 0xFF, sizeof(sectorbuf));
-        } else {
-            if (ReadCartSave(sectorbuf, ent.phys_sec * 0x1000, sizeof(sectorbuf), cdata) != 0)
-                return 1;
-        }
-        
-        if (ctr_cdata->save_crypto_repeating_ctr) {
-            for (u32 cursect = 0; cursect < 8; cursect++) {
-                memset(ctr, 0, sizeof(ctr));
-                ctr_decrypt_byte(&sectorbuf[0x200 * cursect], &sectorbuf[0x200 * cursect], 0x200, 0, AES_CNT_CART_SAVE_MODE, ctr);
-            }
-        } else {
-            ctr_decrypt_byte(sectorbuf, sectorbuf, sizeof(sectorbuf), cur_sector * 0x1000, AES_CNT_CART_SAVE_MODE, ctr);
-        }
-        
-        memcpy(&buffer[outbuf_offset], &sectorbuf[in_sector_start], chunksize);
-        outbuf_offset += chunksize;
-        count -= chunksize;
-        offset += chunksize;
-    }
-    
-    return 0;
-}
-
-static u32 ReadDecryptedCard2Save(u8 *buffer, u64 offset, u64 count, CartData* cdata) {
-    if (offset >= cdata->save_size) return 1;
-    if (offset + count > cdata->save_size) count = cdata->save_size - offset;
-    
-    CartDataCtr *ctr_cdata = (CartDataCtr *)cdata;
-
-    u32 first_sector = offset / 0x200;
-    u32 outbuf_offset = 0;
-    
-    u8 ctr[16];
-    memset(ctr, 0, sizeof(ctr));
-
-    u32 crypto_keyslot = ctr_cdata->save_crypto_type == CARD_SAVE_CRYPTO_V1_N3DS ? CARD_SAVE_CRYPTO_KEYSLOT_N3DS : CARD_SAVE_CRYPTO_KEYSLOT_O3DS;
-    use_aeskey(crypto_keyslot);
-
-    u8 sector_tmp[0x200];
-    
-    // handle misalignment at the start
-    u32 in_first_sector_offset = offset % 0x200;
-    if (in_first_sector_offset) {
-        u32 sector_remain = 0x200 - in_first_sector_offset;
-        u32 misalignsize = min(sector_remain, count);
-        
-        if (ReadCartSave(sector_tmp, first_sector * 0x200, sizeof(sector_tmp), cdata) != 0)
-            return 1;
-        
-        ctr_decrypt_byte(sector_tmp, sector_tmp, sizeof(sector_tmp), first_sector * 0x200, AES_CNT_CART_SAVE_MODE, ctr);
-        memcpy(buffer, &sector_tmp[in_first_sector_offset], misalignsize);
-        outbuf_offset += misalignsize;
-        offset += misalignsize;
-        count -= misalignsize;
-    }
-    
-    // offset is now aligned, size may still not be, though
-    u32 aligned_size = count & ~(0x200-1);
-    if (aligned_size) {
-        if (ReadCartSave(&buffer[outbuf_offset], offset, aligned_size, cdata) != 0)
-            return 1;
-        
-        ctr_decrypt_byte(&buffer[outbuf_offset], &buffer[outbuf_offset], aligned_size, offset, AES_CNT_CART_SAVE_MODE, ctr);
-        
-        count -= aligned_size;
-        offset += aligned_size;
-        outbuf_offset += aligned_size;
-    }
-    
-    // only ending misalignment remains, if applicable
-    if (count) {
-        if (ReadCartSave(sector_tmp, offset, sizeof(sector_tmp), cdata) != 0)
-            return 1;
-        
-        ctr_decrypt_byte(sector_tmp, sector_tmp, sizeof(sector_tmp), offset, AES_CNT_CART_SAVE_MODE, ctr);
-        
-        memcpy(&buffer[outbuf_offset], &sector_tmp, count);
-    }
-    
-    return 0;
-}
-
-u32 ReadDecryptedCartSave(u8* buffer, u64 offset, u64 count, CartData* cdata) {
-    switch (cdata->save_type) {
-        case CARD_SAVE_SPI:
-            return ReadDecryptedCard1Save(buffer, offset, count, cdata);
-            break;
-        
-        case CARD_SAVE_CARD2:
-            return ReadDecryptedCard2Save(buffer, offset, count, cdata);
-            break;
-        
-        default:
-            return 1;
-    }
-}
-
 u32 WriteCartSave(const u8* buffer, u64 offset, u64 count, CartData* cdata) {
     if (offset >= cdata->save_size) return 1;
     if (offset + count > cdata->save_size) count = cdata->save_size - offset;
-    int res = CardSPIWriteSaveData(cdata->spi_save_type, offset, buffer, count);
-    if (cdata->cart_type & CART_CTR) {
-        CartDataCtr *ctr_cdata = (CartDataCtr *)cdata;
-        if (ctr_cdata->wear_leveling.type != CARD_SAVE_WEAR_LEVELING_NONE) {
-            InitCtrCardSaveWearLeveling(cdata); // reload wear-leveling data, might've been invalidated by user
-            if (*GetMountPath() && !ctr_cdata->wear_leveling.initialized && (DriveType(GetMountPath()) & DRV_CART) && strstr(GetMountPath(), ".sav")) {
-                // unmount the virtual DISA archive if user invalidated the encrypted and wear-leveled source
-                InitImgFS(NULL);
-            }
-        }
-    }
-    return res == 0 ? 0 : 1;
+    return (CardSPIWriteSaveData(cdata->spi_save_type, offset, buffer, count) == 0) ? 0 : 1;
 }
