+++ arm9/source/gamecart/save_ctr.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "gamecart.h"
+
+int InitCtrCardSave(CartData *);
+u32 ReadDecryptedCtrCardSave(u8* buffer, u64 offset, u64 count, CartData* cdata);
