+++ arm9/source/virtual/vcart.c
@@ -1,6 +1,9 @@
 #include "vcart.h"
+#include "fsdrive.h"
+#include "fsinit.h"
 #include "gamecart.h"
-#include "ui.h"
+#include "image.h"
+#include "save_ctr.h"
 
 #define FAT_LIMIT   0x100000000
 #define VFLAG_DECRYPTED_SAVEGAME (1UL<<27)
@@ -12,6 +15,7 @@
 static CartData* cdata = NULL;
 static bool cart_init = false;
 static bool cart_checked = false;
+static bool enable_dec_ctr_save = false;
 
 u32 InitVCartDrive(void) {
     if (!cart_checked) cart_checked = true;
@@ -20,6 +24,10 @@ u32 InitVCartDrive(void) {
     if (!cart_init && cdata) {
         free(cdata);
         cdata = NULL;
+    }
+    if (cart_init && (cdata->cart_type & CART_CTR)) {
+        // for compatibility purposes save crypto and wear leveling init are optional
+        enable_dec_ctr_save = InitCtrCardSave(cdata) == 0;
     }
     return cart_init ? cdata->cart_id : 0;
 }
@@ -35,10 +43,6 @@ bool ReadVCartDir(VirtualFile* vfile, VirtualDir* vdir) {
     memset(vfile, 0, sizeof(VirtualFile));
     vfile->keyslot = 0xFF; // unused
     vfile->flags = VFLAG_READONLY;
-    
-    CartDataCtr *ctr_cdata = (CartDataCtr *)cdata;
-    bool ctr_dec_save_supported = (cdata->cart_type & CART_CTR) && ctr_cdata->save_crypto_type != CARD_SAVE_CRYPTO_INVALID &&
-        (ctr_cdata->wear_leveling.type == CARD_SAVE_WEAR_LEVELING_NONE || ctr_cdata->wear_leveling.initialized);
 
     while (++vdir->index <= 10) {
         if ((vdir->index == 0) && (cdata->data_size < FAT_LIMIT)) { // standard full rom
@@ -79,7 +83,7 @@ bool ReadVCartDir(VirtualFile* vfile, VirtualDir* vdir) {
                 vfile->flags |= VFLAG_READONLY;
             }
             return true;
-        } else if ((vdir->index == 8) && ctr_dec_save_supported) {
+        } else if ((vdir->index == 8) && enable_dec_ctr_save) {
             snprintf(vfile->name, 32, "%s.dec.sav", name);
             vfile->size = (cdata->cart_id & 0x8000000) /* card2 */ ? cdata->save_size : cdata->save_size - 0x1000;
             vfile->flags = VFLAG_DECRYPTED_SAVEGAME | VFLAG_READONLY /* for now */;
@@ -106,7 +110,7 @@ int ReadVCartFile(const VirtualFile* vfile, void* buffer, u64 offset, u64 count)
     else if (vfile->flags & VFLAG_SAVEGAME)
         return ReadCartSave(buffer, foffset, count, cdata);
     else if (vfile->flags & VFLAG_DECRYPTED_SAVEGAME)
-        return ReadDecryptedCartSave(buffer, foffset, count, cdata);
+        return ReadDecryptedCtrCardSave(buffer, foffset, count, cdata);
     else if (vfile->flags & VFLAG_GAMECART_NFO)
         return ReadCartInfo(buffer, foffset, count, cdata);
 
@@ -117,7 +121,15 @@ int ReadVCartFile(const VirtualFile* vfile, void* buffer, u64 offset, u64 count)
 int WriteVCartFile(const VirtualFile* vfile, const void* buffer, u64 offset, u64 count) {
     if (!cdata) return -1;
     if (vfile->flags & VFLAG_SAVEGAME) {
-        return WriteCartSave(buffer, offset, count, cdata);
+        int res = WriteCartSave(buffer, offset, count, cdata);
+        if (cdata->cart_type & CART_CTR) {
+            enable_dec_ctr_save = InitCtrCardSave(cdata) == 0;
+            if (*GetMountPath() && !enable_dec_ctr_save && (DriveType(GetMountPath()) & DRV_CART) && strstr(GetMountPath(), ".sav")) {
+                // unmount the virtual DISA archive if user invalidated the encrypted and wear-leveled source
+                InitImgFS(NULL);
+            }
+        }
+        return res;
     }
     return -1;
 }
